refactored command registration
